+++ src/main/java/com/wanted/needswork/repository/JobSeekerRepository.java
@@ -4,8 +4,10 @@ import com.wanted.needswork.models.JobSeeker;
 import com.wanted.needswork.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
+
 public interface JobSeekerRepository extends JpaRepository <JobSeeker, Integer> {
 
-    JobSeeker getJobSeekerByUserId(Integer userId);
+    JobSeeker getJobSeekerByUserId(BigInteger userId);
 }
